[inner-2101] integration sql_log\show trace\show @@sql.xx\show @@cost_time

+++ src/main/java/com/actiontech/dble/statistic/sql/handler/AssociateTablesByEntryByUserCalcHandler.java
@@ -54,7 +54,7 @@ public class AssociateTablesByEntryByUserCalcHandler implements StatisticDataHan
                             checkEliminate();
                             currRecord = new Record(fEntry.getFrontend().getUserId(), fEntry.getFrontend(), tables);
                         }
-                        currRecord.addSelect(fEntry.getExaminedRows().longValue(), fEntry.getRows(), fEntry.getDuration());
+                        currRecord.addSelect(fEntry.getExaminedRows(), fEntry.getRows(), fEntry.getDuration());
                         if (isNew) {
                             records.put(key, currRecord);
                         }
